ClipPath to svg (#5234)
* svg plain examples works

* more code reuse in visual tests
