+++ test/node_test_setup.js
@@ -6,10 +6,13 @@ global.fs = require('fs');
 global.visualCallback = {
   addArguments: function() {},
 };
+global.visualTestLoop = require('./lib/visualTestLoop').visualTestLoop;
+global.getFixture = require('./lib/visualTestLoop').getFixture;
+global.getAsset = require('./lib/visualTestLoop').getAsset;
 global.imageDataToChalk = function(imageData) {
   // actually this does not work on travis-ci, so commenting it out
   return '';
-  var block = String.fromCharCode(9608)
+  var block = String.fromCharCode(9608);
   var data = imageData.data;
   var width = imageData.width;
   var height = imageData.height;
